Fixed the crash on 32-bit operating systems

+++ src/globals.h
@@ -838,7 +838,7 @@ struct WorldLevel_t
 //    Visible As Boolean 'true if it should be shown on the map
     bool Visible = false;
 //End Type
-    long long Z = 0;
+    int64_t Z = 0;
     int index = 0;
 };
 
@@ -905,7 +905,7 @@ struct Tile_t
 //    Type As Integer
     int Type = 0;
 //End Type
-    long long Z = 0;
+    int64_t Z = 0;
     bool Active = true;
 };
 
@@ -919,7 +919,7 @@ struct Scene_t
 //    Active As Boolean 'if false this won't be shown. used for paths that become available on a scene
     bool Active = false;
 //End Type
-    long long Z = 0;
+    int64_t Z = 0;
 };
 
 //Public Type WorldPath 'World Paths
@@ -932,7 +932,7 @@ struct WorldPath_t
 //    Type As Integer
     int Type = 0;
 //End Type
-    long long Z = 0;
+    int64_t Z = 0;
     int index = 0;
 };
 
@@ -946,7 +946,7 @@ struct WorldMusic_t
 //    EXTRA: Custom Music
     std::string MusicFile;
 //End Type
-    long long Z = 0;
+    int64_t Z = 0;
     bool Active = true;
 };
